requirements.txt
# ================================================================
# Requirements - Experimento GBML Concept Drift
# Python 3.8+
# ================================================================

# --- Machine Learning & Data Science ---
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# --- River (Streaming ML) ---
river>=0.18.0

# --- Plotting & Visualization ---
matplotlib>=3.4.0
seaborn>=0.11.0

# --- YAML Configuration ---
PyYAML>=6.0

# --- String Similarity (para rule_diff_analyzer.py) ---
python-Levenshtein>=0.12.0
# Alternativa: editdistance>=0.5.3

# --- Multiprocessing (já incluído no Python, mas explícito) ---
# multiprocessing - built-in

# --- Utilities ---
tqdm>=4.62.0  # Progress bars (opcional, mas útil)

# --- Testing (opcional) ---
# pytest>=7.0.0
# pytest-cov>=3.0.0

# ================================================================
# NOTAS:
# - Instalação: pip install -r requirements.txt
# - No Colab, algumas libs já estão pré-instaladas
# - python-Levenshtein pode precisar de compilador C
#   Se falhar, usar: pip install python-Levenshtein-wheels
# ================================================================
